[changed] Disable native autocomplete (#93)
You can do this via `inputProps`, but since the whole point of this component
is to provide a custom autocomplete solution it makes sense to disable the
native counterpart. If this proves to be problematic we can change the ordering
so that the consumer can override it via `inputProps: { autoComplete: 'on' }`.